+++ src/components/attribution-editor.js
@@ -16,7 +16,7 @@ import CreditAdd from './credit-add'
 import TextareaAutosize from './textarea-autosize'
 
 import blockMetaSchema from '../schema/block-meta'
-import {isFileEvent} from '../util/drop'
+import {isDragFileEvent, isDropFileEvent} from '../util/drop'
 
 
 class AttributionEditor extends React.Component {
@@ -24,13 +24,17 @@ class AttributionEditor extends React.Component {
     super(props)
     this.state =
       { block: props.initialBlock
+      , showDropIndicator: false
       }
 
     this.boundOnDragOver = this.onDragOver.bind(this)
+    this.boundOnDragEnter = this.onDragEnter.bind(this)
+    this.boundOnDragLeave = this.onDragLeave.bind(this)
     this.boundOnDrop = this.onDrop.bind(this)
     this.boundOnChange = this.onChange.bind(this)
     this.boundOnMoreClick = this.onMoreClick.bind(this)
     this.boundOnUploadRequest = this.onUploadRequest.bind(this)
+    this.boundOnCoverRemove = this.onCoverRemove.bind(this)
   }
   componentWillReceiveProps (props) {
     this.setState({block: props.initialBlock})
@@ -40,7 +44,7 @@ class AttributionEditor extends React.Component {
     const {type, metadata, cover} = block
     const schema = blockMetaSchema[type] || blockMetaSchema.default
 
-    const menus = renderMenus(type, schema, metadata, cover, this.boundOnChange, this.boundOnMoreClick, this.boundOnUploadRequest)
+    const menus = renderMenus(type, schema, metadata, cover, this.boundOnChange, this.boundOnMoreClick, this.boundOnUploadRequest, this.boundOnCoverRemove)
 
     return el('div'
       , { className: 'AttributionEditor'
@@ -52,6 +56,8 @@ class AttributionEditor extends React.Component {
           , position: 'relative'
           }
         , onDragOver: this.boundOnDragOver
+        , onDragEnter: this.boundOnDragEnter
+        , onDragLeave: this.boundOnDragLeave
         , onDrop: this.boundOnDrop
         }
       , this.renderCover()
@@ -75,6 +81,7 @@ class AttributionEditor extends React.Component {
           , el(DropdownGroup, {menus})
         )
       )
+      , this.renderDropIndicator()
       , el('div'
         , { style: {clear: 'both'} }
       )
@@ -128,7 +135,7 @@ class AttributionEditor extends React.Component {
     , 'We were unable to find the image originally saved with this block.'
     , el(Space, {auto: true})
     , el(Button
-      , { onClick: () => this.onUploadRequest()
+      , { onClick: this.boundOnUploadRequest
         , rounded: true
         , color: 'error'
         , backgroundColor: 'white'
@@ -151,10 +158,39 @@ class AttributionEditor extends React.Component {
       }
     )
   }
+  renderDropIndicator () {
+    const {showDropIndicator} = this.state
+    if (!showDropIndicator) return
+
+    return el('div'
+    , { style:
+        { position: 'absolute'
+        , top: 0
+        , left: 0
+        , width: '100%'
+        , height: '100%'
+        , textAlign: 'center'
+        , fontSize: 36
+        , paddingTop: '5%'
+        , backgroundColor: 'rgba(255, 255, 255, 0.75)'
+        , border: '5px black dashed'
+        }
+      }
+    , 'DROP TO REPLACE THIS IMAGE'
+    )
+  }
   onUploadRequest () {
     const {store} = this.context
     const {id} = this.props
     store.routeChange('MEDIA_BLOCK_REQUEST_COVER_UPLOAD', id)
+  }
+  onCoverRemove () {
+    const {store} = this.context
+    const {id} = this.props
+    // Send change up to store
+    const block = store.routeChange('MEDIA_BLOCK_COVER_REMOVE', id)
+    // Send change to view
+    this.setState({block})
   }
   onChange (path, value) {
     const {store} = this.context
@@ -165,13 +201,37 @@ class AttributionEditor extends React.Component {
     this.setState({block})
   }
   onDragOver (event) {
-    if (event.dataTransfer.types[0] !== 'Files') return
     event.preventDefault()
   }
-  onDrop (event) {
-    if (!isFileEvent(event)) return
-    if (!this.canChangeCover()) return
+  onDragEnter (event) {
+    if (!isDragFileEvent(event)) return
     event.preventDefault()
+    const {showDropIndicator} = this.state
+    if (showDropIndicator) return
+    if (!this.canChangeCover()) return
+    this.setState({showDropIndicator: true})
+  }
+  onDragLeave (event) {
+    event.preventDefault()
+    const {showDropIndicator} = this.state
+    if (!showDropIndicator) return
+    // HACK
+    // Check if actually left as opposed to drag over child
+    const x = event.clientX + window.scrollX
+    const y = event.clientY + window.scrollY
+    const {offsetTop, offsetHeight, offsetLeft, offsetWidth} = event.currentTarget.parentNode
+    const top = offsetTop
+    const bottom = top + offsetHeight
+    const left = offsetLeft
+    const right = left + offsetWidth
+    if (y <= (top + 50) || y >= bottom || x <= left || x >= right) {
+      this.setState({showDropIndicator: false})
+    }
+  }
+  onDrop (event) {
+    if (!isDropFileEvent(event)) return
+    event.preventDefault()
+    if (!this.canChangeCover()) return
     event.stopPropagation()
 
     const {store} = this.context
@@ -181,6 +241,7 @@ class AttributionEditor extends React.Component {
       , file: event.dataTransfer.files[0]
       }
     )
+    this.setState({showDropIndicator: false})
   }
   onMoreClick (key) {
     const {store} = this.context
@@ -270,7 +331,7 @@ function renderTextField (key, label, value, onChange) {
   )
 }
 
-function renderMenus (type, schema, metadata = {}, cover, onChange, onMoreClick, onUploadRequest) {
+function renderMenus (type, schema, metadata = {}, cover, onChange, onMoreClick, onUploadRequest, onCoverRemove) {
   let menus = []
   if (schema.isBasedOnUrl && metadata.isBasedOnUrl != null) {
     menus.push(
@@ -296,8 +357,9 @@ function renderMenus (type, schema, metadata = {}, cover, onChange, onMoreClick,
   if (cover || schema.changeCover) {
     const hasCover = (cover != null)
     const allowCoverChange = schema.changeCover
+    const allowCoverRemove = (cover && schema.removeCover)
     menus.push(
-      renderImageEditor(hasCover, allowCoverChange, type, metadata.title, metadata.coverPrefs, onChange, onUploadRequest)
+      renderImageEditor(hasCover, allowCoverChange, allowCoverRemove, type, metadata.title, metadata.coverPrefs, onChange, onUploadRequest, onCoverRemove)
     )
   }
   menus.push(
@@ -327,17 +389,19 @@ function renderCreditEditor (onlyUrl, key, label, item, onChange, path) {
   )
 }
 
-function renderImageEditor (hasCover, allowCoverChange, type, title, coverPrefs = {}, onChange, onUploadRequest) {
+function renderImageEditor (hasCover, allowCoverChange, allowCoverRemove, type, title, coverPrefs = {}, onChange, onUploadRequest, onCoverRemove) {
   const {filter, crop, overlay} = coverPrefs
   return el(ImageEditor
   , { hasCover
     , allowCoverChange
+    , allowCoverRemove
     , title
     , filter
     , crop
     , overlay
     , onChange
     , onUploadRequest
+    , onCoverRemove
     , type
     , name: 'Image'
     , label: 'Image'