+++ CHANGES.md
@@ -1,5 +1,12 @@
 ## dev
 
+## 0.17.4 - 2016-06-23
+
+* NEW -- Images can be removed from `article` blocks (#243)
+* Less-chunky metadata editing on small screens (#236)
+* Disable ProseMirror format menu on media block selection
+* Indicator to drop file on block (#235)
+
 ## 0.17.3 - 2016-06-22
 
 * UI in menus add code and location blocks (#245)